+++ src/constants.js
@@ -1,9 +1,10 @@
-export const VERSION = '0.2.0-alpha.1';
+export const VERSION = '0.2.0-alpha.2';
 
 export const SKIPPED_DIRECTORIES = new Set([
   '.git',
   '.hg',
   '.svn',
+  '.setuplens',
   '.next',
   '.npm-cache',
   '.nuxt',